requirements.txt
python-telegram-bot==20.4
requests==2.28.2
python-dotenv==1.0.0